+++ components/GraphicProject/graphicElasticMock.js
@@ -17,183 +17,183 @@ export const graphicElasticMock = [
         {
           title: 'Mon titre graphique',
           visibility: 'PUBLIC',
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_FRENCH_Twitter_Output.png',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_FRENCH_Twitter_Output.png',
           language: {
             language_code: 'fr',
             text_direction: 'ltr',
             locale: 'fr-fr',
             display_name: 'French',
-            native_name: 'French'
+            native_name: 'French',
           },
-          name: 'CAPTIONS_Made_in_America_FRENCH_Twitter_Output.png',
+          filename: 'CAPTIONS_Made_in_America_FRENCH_Twitter_Output.png',
           alt: 'Texte alternatif français',
           filesize: '600 MB',
           width: 800,
           height: 600,
-          social: 'Twitter',
-          style: 'Quote'
+          social: ['Twitter'],
+          style: 'Quote',
         },
         {
           title: 'Mon titre graphique FACEBOOK',
           visibility: 'PUBLIC',
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_FRENCH_Facebook_Output.png',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_FRENCH_Facebook_Output.png',
           language: {
             language_code: 'fr',
             text_direction: 'ltr',
             locale: 'fr-fr',
             display_name: 'French',
-            native_name: 'French'
+            native_name: 'French',
           },
-          name: 'CAPTIONS_Made_in_America_FRENCH_Facebook_Output.png',
+          filename: 'CAPTIONS_Made_in_America_FRENCH_Facebook_Output.png',
           alt: 'Texte alternatif français',
           filesize: '600 MB',
           width: 800,
           height: 600,
-          social: 'Facebook',
-          style: 'Quote'
+          social: ['Facebook', 'Instagram'],
+          style: 'Quote',
         },
         {
           title: 'My graphic title',
           visibility: 'PUBLIC',
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.png',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.png',
           language: {
             language_code: 'en',
             text_direction: 'ltr',
             locale: 'en-us',
             display_name: 'English',
-            native_name: 'English'
+            native_name: 'English',
           },
-          name: 'CAPTIONS_Made_in_America_English_Output.png',
+          filename: 'CAPTIONS_Made_in_America_English_Output.png',
           alt: 'English alt text',
           filesize: '600 MB',
           width: 800,
           height: 600,
-          social: 'Twitter',
-          style: 'Quote'
-        }
+          social: ['Twitter'],
+          style: 'Quote',
+        },
       ],
       supportFiles: [
         {
           visibility: 'INTERNAL',
           editable: true,
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.psd',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.psd',
           language: {
             language_code: 'en',
             text_direction: 'ltr',
             locale: 'en-us',
             display_name: 'English',
-            native_name: 'English'
-          }
+            native_name: 'English',
+          },
         },
         {
           visibility: 'INTERNAL',
           editable: true,
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.ai',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.ai',
           language: {
             language_code: 'en',
             text_direction: 'ltr',
             locale: 'en-us',
             display_name: 'English',
-            native_name: 'English'
-          }
+            native_name: 'English',
+          },
         },
         {
           visibility: 'INTERNAL',
           editable: true,
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.png',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.png',
           language: {
             language_code: 'en',
             text_direction: 'ltr',
             locale: 'en-us',
             display_name: 'English',
-            native_name: 'English'
-          }
+            native_name: 'English',
+          },
         },
         {
           visibility: 'INTERNAL',
           editable: true,
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_French_Output.psd',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_French_Output.psd',
           language: {
             language_code: 'fr',
             text_direction: 'ltr',
             locale: 'fr-fr',
             display_name: 'French',
-            native_name: 'French'
-          }
+            native_name: 'French',
+          },
         },
         {
           visibility: 'INTERNAL',
           editable: true,
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_French_Output.ai',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_French_Output.ai',
           language: {
             language_code: 'fr',
             text_direction: 'ltr',
             locale: 'fr-fr',
             display_name: 'French',
-            native_name: 'French'
-          }
+            native_name: 'French',
+          },
         },
         {
           visibility: 'INTERNAL',
           editable: false,
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_French_Output.doc',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_French_Output.doc',
           language: {
             language_code: 'fr',
             text_direction: 'ltr',
             locale: 'fr-fr',
             display_name: 'French',
-            native_name: 'French'
-          }
+            native_name: 'French',
+          },
         },
         {
           visibility: 'INTERNAL',
           editable: false,
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_French_Output.pdf',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_French_Output.pdf',
           language: {
             language_code: 'fr',
             text_direction: 'ltr',
             locale: 'fr-fr',
             display_name: 'French',
-            native_name: 'French'
-          }
+            native_name: 'French',
+          },
         },
         {
           visibility: 'INTERNAL',
           editable: false,
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.doc',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.doc',
           language: {
             language_code: 'en',
             text_direction: 'ltr',
             locale: 'en-us',
             display_name: 'English',
-            native_name: 'English'
-          }
+            native_name: 'English',
+          },
         },
         {
           visibility: 'INTERNAL',
           editable: false,
-          srcUrl: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.pdf',
+          url: 'https://cdp-video-tst.s3.amazonaws.com/2020/04/commons.america.gov_ck8h90w6x030l0765onryiqdt/CAPTIONS_Made_in_America_English_Output.pdf',
           language: {
             language_code: 'en',
             text_direction: 'ltr',
             locale: 'en-us',
             display_name: 'English',
-            native_name: 'English'
-          }
-        }
+            native_name: 'English',
+          },
+        },
       ],
       tags: [
         {
           name: 'biomedical science',
-          id: 'KBb5K3EB-L8EaltkVCV8'
-        }
+          id: 'KBb5K3EB-L8EaltkVCV8',
+        },
       ],
       categories: [
         {
           name: 'science & technology',
-          id: 'TBb5K3EB-L8EaltkViXQ'
-        }
-      ]
-    }
-  }
+          id: 'TBb5K3EB-L8EaltkViXQ',
+        },
+      ],
+    },
+  },
 ];