+++ components/popups/Popup.js
@@ -3,16 +3,16 @@ import PropTypes from 'prop-types';
 import { Header } from 'semantic-ui-react';
 import './Popup.scss';
 
-const Popup = props => (
+const Popup = ( { title, children } ) => (
   <div>
-    <Header as="h2">{ props.title }</Header>
-    { props.children }
+    <Header as="h2">{ title }</Header>
+    { children }
   </div>
 );
 
 Popup.propTypes = {
   title: PropTypes.string,
-  children: PropTypes.node
+  children: PropTypes.node,
 };
 
 export default Popup;